requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rapidfuzz>=3.0.0
loguru>=0.7.0
typer>=0.12.0
rich>=13.0.0
pyyaml>=6.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
anyio>=4.0.0
